+++ modules/core/src/directive/dynamic-template.directive.ts
@@ -7,6 +7,7 @@ import { Directive, Input, TemplateRef } from "@angular/core";
 export class DynamicTemplateDirective {
 
     @Input() modelId: string;
+    @Input() type: string | null = null;
 
     constructor(public templateRef: TemplateRef<any>) {}
 }